PLAT-229: add codecov.io integration (#78)
Add codecov.io integration. The per-line diffs on codecov.io work kind of weird. Ilya Ozherelyev and I believe it doesn't like that there is no code coverage info for master branch. We decided to merge this PR and see it will solve the issue. The PR doesn't make anything worse thus it's OK to solve the issue in a separate PR.